Worried enough to lose progress to upload
In very early development - so far the function can detect clusters via hdbscan from a .fit file, generate a formula for the ellipse of best fit, and from there calculate and generate a numpy array of coordinates at ever meter of the track of best fit. Still need to add a track detection function (decide based on ellipse size and fit quality which clusters to ignore), snap the existing activity stream points to the track points, and reencode the fit file for upload to Strava.